+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog
 2.0 (unreleased)
 ----------------
 
+- Remove Timezone field from IEventBasic behavior. Instead, store timezone
+  information directly in the tzinfo object on start and end datetime objects.
+  [yenzenz, thet]
+
 - Remove Archetypes subpackage.
   [yenzenz, thet]